+++ pulse2percept/utils/__init__.py
@@ -17,9 +17,7 @@
 """
 from .base import (PrettyPrint, FreezeError, Frozen, Data, bijective26_name,
                    cached, gamma)
-from .geometry import (Grid2D, VisualFieldMap, RetinalMap, CorticalMap,
-                       Curcio1990Map, Watson2014Map, Watson2014DisplaceMap,
-                       cart2pol, pol2cart, delta_angle)
+from .geometry import (cart2pol, pol2cart, delta_angle)
 from .array import is_strictly_increasing, radial_mask, sample, unique
 from .images import center_image, scale_image, shift_image, trim_image
 from .convolution import center_vector, conv
@@ -37,14 +35,12 @@ __all__ = [
     'center_vector',
     'circ_r2_score',
     'conv',
-    'Curcio1990Map',
     'Data',
     'delta_angle',
     'deprecated',
     'FreezeError',
     'Frozen',
     'gamma',
-    'Grid2D',
     'is_strictly_increasing',
     'parfor',
     'pol2cart',
@@ -56,9 +52,4 @@ __all__ = [
     'shift_image',
     'trim_image',
     'unique',
-    'VisualFieldMap',
-    'RetinalMap',
-    'CorticalMap',
-    'Watson2014DisplaceMap',
-    'Watson2014Map'
 ]
